requirements.txt
lyricsgenius==1.8.4
pygame==1.9.6
python-dotenv==0.13.0
spotipy==2.12.0
